変更 - SuspenseとErrorBoundaryでエラーをCatchしエラー画面を出す 同階層にキャッチとHookがあったため、Hook側を別ファイルに書き出し、親子関係にすることで実装する
